+++ src/lib/error.model.ts
@@ -1,0 +1,9 @@
+export interface NonfatalError {
+    title: string,
+    description: string,
+    error: {
+        code: string | number,
+        message: string,
+        stack: string
+    }
+}