+++ src/renderer/components/index.ts
@@ -2,3 +2,4 @@ export * from './main.component';
 export * from './settings.component';
 export * from './home.component';
 export * from './steam-devices.component';
+export * from './error.component';