+++ src/renderer/main.routing.ts
@@ -1,7 +1,7 @@
 import { ModuleWithProviders } from '@angular/core';
 import { Routes as AngularRoutes, RouterModule } from '@angular/router';
 
-import { HomeComponent, SettingsComponent, SteamDevicesComponent } from './components';
+import { HomeComponent, SettingsComponent, SteamDevicesComponent, ErrorComponent } from './components';
 
 const AppRouter: AngularRoutes = [
     {
@@ -11,6 +11,10 @@ const AppRouter: AngularRoutes = [
     {
         path: 'settings',
         component: SettingsComponent
+    },
+    {
+        path: 'error',
+        component: ErrorComponent
     },
     {
         path: 'steam-devices',