+++ src/renderer/services/index.ts
@@ -1,3 +1,4 @@
 export * from './settings.service';
 export * from './steam-devices.service';
 export * from './window.service';
+export * from './error.service';